♻️ refactor: Report 도메인 FacadeService 관련 코드 수정 및 리팩토링 진행
Ref: #TRB-9
